refactor: follow up w/ import update
